Use environment variables in Windows paths
This better mimics the Unix like `~` usage for the others, removing the need for "variable replacement" by the user. Paths in this format are supported in a lot of places in Windows, such as File Explorer, the classic Run dialog, and traditional CMD.EXE. PowerShell users should already know how to convert these already, so not including that here. (For reference: `%USERPROFILE%` is supported as `~` in PowerShell, and `$env:LocalAppData` is the variable format for `%LocalAppData%`.)