Add more detail to the MacOS install instructions
I just setup the ActivityWatch on MacOS and I ran into the other issues outlined in #334.  I want to update the note to try to make it more clear that a new user still needs to work around this issue, bc Catalina was old I figured this was already taken care of and didn't get re-directed to the issue until google searching.