fix GNOME desktop entry instructions
Gotta use absolute paths in `Exec` in `.desktop` files or they don't show up automatically. Not sure if this is true for everyone/always, but it was for me.

Sources:

https://bbs.archlinux.org/viewtopic.php?id=256867

https://discourse.gnome.org/t/application-entries-desktop-not-displayed/3670

https://discussion.fedoraproject.org/t/how-to-create-a-desktop-file-to-show-up-among-my-apps/128596/5

These sources indicate the underlying issue appears has to do with an issue with the user's non-interactive path, but I think the issue might actually be the tilde is not expanded. I didn't attempt to debug further, I just used an absolute path and it worked. Seems like the safer (and more annoying) way to go to make it work for most people. I do with tilde worked more reliably.

FFR, my desktop right now is GNOME, as shipped with 64-bir Ubuntu 22.04 LTS. I think that means GNOME 3 and Wayland.